Localstorage items removed after loading the page
To Vajid : You have changed the API to check available seats.
In kafka backend, in Payment Model, in complete_payment, { payment_successful : true } is not working and even data is not going into the database. So not getting the true value at the front-end and page is not working.
Temporary page is working with (true)
Check it Once.